Entrenamiento vet con cuda/requirements.txt
ultralytics>=8.0.0
opencv-python
pyyaml
tqdm
numpy
matplotlib
pandas
seaborn
tensorboard
torch>=2.0.0
